# docs/swagger.yaml
basePath: /v1
definitions:
  bahirehasab.Basic:
    properties:
      abektie:
        description: Abekti (አበቅቴ)
        type: integer
      beale_metiq:
        description: BealeMetiq (በዓለ-መጥቅ), 1 =መስከረም፣ 2 = ጥቅምት
        type: integer
      mebaja_hamer:
        description: MebajaHamer (መባጃ ሃመር)
        type: integer
      medeb:
        description: Medeb (መደብ)
        type: integer
      metiq:
        description: Metiq (መጥቅ)
        type: integer
      nenewie:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Nenewie የነነዌ ፆም የሚውልበት ቀን
      wenber:
        description: Wenber (ወንበር)
        type: integer
    type: object
  bahirehasab.Date:
    properties:
      date_of_the_month:
        type: integer
      month_of_the_year:
        type: integer
    type: object
  bahirehasab.Fasting:
    properties:
      abiy:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Abiy tsome (አብይ ፆም)
      debrezeit:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: DebreZeit ደብረ ዘይት
      dihnet:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Dihnet (ፆመ ድህነት), Wednesday and Friday Weekly fast
      erget:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Erget (እርገት)፣ Ascension
      hawariyat:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Hawariyat ( ፆመ ሐዋሪያት), fast of the Holy Apostles
      hosanna:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Hosanna ሆሳህና
      peraklitos:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Peraklitos (ጰራቅሊጦስ)፣ Paraclete
      rkbekahnat:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: RkbeKahnat (ርክበ ካህናት)
      siklet:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Siklet (ስቅለት), Good Friday
      tinsaye:
        allOf:
        - $ref: '#/definitions/bahirehasab.Date'
        description: Tinsaye (ፋሲካ), resurrection
    type: object
  bahirehasab.Festival:
    properties:
      basic:
        $ref: '#/definitions/bahirehasab.Basic'
      fasting:
        $ref: '#/definitions/bahirehasab.Fasting'
      year:
        $ref: '#/definitions/bahirehasab.Year'
    type: object
  bahirehasab.Year:
    properties:
      day_of_the_week:
        description: DayOfTheWeek the day in which new year fall
        type: string
      evangelist:
        description: Evangelist ወንጌላዊው in name
        type: string
      evangelist_num:
        description: EvangelistNum (Evangelist) ወንጌላዊው in number
        type: integer
      year:
        description: Year in Ethiopian calendar
        type: integer
    type: object
host: calendar.st-son.com
info:
  contact:
    email: support@st-son.com
    name: API Support
    url: https://st-son.com/support
  description: Ethiopian to Gregorian date converter server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://st-son.com/terms/
  title: EthioGrego
  version: "1.0"
paths:
  /:
    get:
      description: Home Page
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Home page
      tags:
      - Date-Converter
  /ad-to-et/{date}:
    get:
      consumes:
      - application/json
      description: Get ethiopian date string from gregorian date
      parameters:
      - description: Date in yy-mm-dd format
        example: '"2024-08-12"'
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Gregorian to Ethiopian
      tags:
      - Date-Conversion
  /bahire-hasab/{year}:
    get:
      consumes:
      - application/json
      description: Get years fasting and festival date
      parameters:
      - description: year
        example: '"2016"'
        in: path
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bahirehasab.Festival'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Bahirehasab (ባህረ-ሐሳብ)
      tags:
      - Bahirehasab
  /et-to-ad/{date}:
    get:
      consumes:
      - application/json
      description: Get gregorian date string from ethiopian date
      parameters:
      - description: Date in yy-mm-dd format
        example: '"2016-12-06"'
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Ethiopian to Gregorian
      tags:
      - Date-Conversion
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"